requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
boto3
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
Django==5.1.3
django-storages
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.54.4
psycopg2-binary
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
